chore: add hook case creator
