+++ packages/rspack-test-tools/src/case/compiler.ts
@@ -16,12 +16,7 @@ export function createCompilerCase(
 	testConfig: string
 ) {
 	const caseConfig: TCompilerCaseConfig = require(testConfig);
-
-	const runner = getSimpleProcessorRunner(src, dist, {
-		it,
-		beforeEach,
-		afterEach
-	});
+	const runner = getSimpleProcessorRunner(src, dist);
 
 	it(caseConfig.description, async () => {
 		await runner(
