+++ packages/rspack-test-tools/src/test/simple.ts
@@ -1,4 +1,4 @@
-import { ITestEnv, ITestProcessor } from "../type";
+import { ITestContext, ITestEnv, ITestProcessor } from "../type";
 import { TestContext } from "./context";
 
 const CONTEXT_MAP: Map<
@@ -9,14 +9,18 @@ const CONTEXT_MAP: Map<
 export function getSimpleProcessorRunner(
 	src: string,
 	dist: string,
-	env: ITestEnv
+	options: {
+		env?: () => ITestEnv;
+		context?: (src: string, dist: string) => ITestContext;
+	} = {}
 ) {
+	const createEnv = options.env || (() => ({ it, beforeEach, afterEach }));
+	const createContext =
+		options.context ||
+		((src: string, dist: string) => new TestContext({ src, dist }));
 	const key = `src: ${src}, dist: ${dist}`;
 	if (!CONTEXT_MAP.has(key)) {
-		const context = new TestContext({
-			src,
-			dist
-		});
+		const context = createContext(src, dist);
 		const runner = async function run(name: string, processor: ITestProcessor) {
 			try {
 				await processor.beforeAll?.(context);
@@ -27,8 +31,8 @@ export function getSimpleProcessorRunner(
 			} catch (e: unknown) {
 				context.emitError(name, e as Error);
 			} finally {
-				await processor.run?.(env, context);
-				await processor.check?.(env, context);
+				await processor.run?.(createEnv(), context);
+				await processor.check?.(createEnv(), context);
 				await processor.after?.(context);
 				await processor.afterAll?.(context);
 			}
